requirements.txt
py-processors==3.0.3
spacy>=3.0.0,<4.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm

fastapi[all]==0.61.1
uvicorn==0.11.8